==> scripts/01_behavior.Rmd <==
---
title: "Behavior Data Analysis"
author: "Rayna M Harris"
date: "Last updated November 22, 2017"
output: md_document
---

```{r setup, echo = F, message=F}
## load libraries 
library(tidyr) ## for respahing data
library(plyr) ## for renmaing factors
library(dplyr) ## for filtering and selecting rows
library(reshape2) ## for melting dataframe
```

This script is use to import, clean, reformat, subset, and summarize data

This first chuck of code loads and wrangles the data.

```{r loaddata}
## read intermediate data (raw data from video tracker program analyzed in matlab)
behavior <- read.csv("../data/fmr1.csv", header = T)
behavior$APA <- as.factor(behavior$APA)
behavior$APA2 <- as.factor(behavior$APA2)

## relevel then rename factors treatment
behavior$APA2 <- factor(behavior$APA2, levels = c("controlconsistent", "consistent", "controlconflict", "conflict"))
levels(behavior$APA2) <- c("yoked-consistent","consistent", "yoked-conflict","conflict")

levels(behavior$Genotype) <- c("WT","FMR1KO") ## relevel genotype

## relevel conflict
behavior$conflict <- ifelse(grepl("conflict", behavior$APA2), "conflict", "consistent") # for splitting
levels(behavior$conflict) <- c("consistent","conflict")

behavior$Time1stEntrLog <- log(behavior$Time1stEntr)  ## log transformation

behavior <- behavior[c(1,3,7,8,10,60,14:59)] # subset data
behavior <- subset(behavior, !is.na(behavior$NumEntrances)) # remove nas

write.csv(behavior, "../results/behaviordata.csv", row.names = F)
```

The videotracking software captures ~ 40 measures about the animal moves around the arena. I also have a handful of columns of meta data that range in their utitlity.
```{r measures}
names(behavior)
```

I focus on these 6 "behavioral" measures with a few descriptive columns about the data. The `ID` identified each unique mouse. I use `TrainSessionCombo` and `TrainSessionCombo` to indicate the order of active place avoiance training sessions. `Genotype` and `APA2` are catagorical variables under investigation. The behaioral measures are explored in detail in the subesent scripts named Fig-Fig7.

```{r overview}

avoidance <- behavior[,c(1,2,8,4,9,12,26:29,15,16)]
summary(avoidance)
```

These group sizes are also shown in Fig 1.

```{r samplesizes}
avoidance %>% 
  filter(TrainSessionCombo == "Retention", Genotype == "WT") %>%
  select(APA2, Genotype)  %>%  summary()
avoidance %>% 
  filter(TrainSessionCombo == "Retention", Genotype == "FMR1KO") %>%
  select(APA2, Genotype)  %>%  summary()
```


This is a long dataframe where `variable` has four levels (shock zone, clockwise, opposite, or counter-clockwise) and `value` has the propportion of time spent in each of those locaitons. 

```{r proptimedataframe}
## proptime
proptime <- behavior[,c(1,2,4,8,9,12,26:29)]
proptime <- melt(proptime, id.vars = c("ID", "Genotype", "TrainSession",
                                       "APA2", "TrainSessionCombo", "TrainSessionComboNum")) 

levels(proptime$variable)
write.csv(proptime, "../results/behaviorproptime.csv", row.names = F)
```

This chunck of code is for summarizing the data to plot mean +/- SEM because box plots are too unwieldly for this type of data. I add a `measure` column for facetting or subsetting the data.

```{r summarized}
numentr <- dplyr::summarise(group_by(behavior, Genotype, APA2, TrainSessionComboNum, conflict), 
                            m = mean(NumEntrances), 
                            se = sd(NumEntrances)/sqrt(length(NumEntrances)), 
                            len = length(NumEntrances))
fivenum(numentr$m)

pathentr <- dplyr::summarise(group_by(behavior, Genotype, APA2, TrainSessionComboNum, conflict), 
                             m = mean(Path1stEntr), 
                             se = sd(Path1stEntr)/sqrt(length(Path1stEntr)), 
                             len = length(Path1stEntr))
fivenum(pathentr$m)

numentr$measure <- "Number of Entrances"
pathentr$measure <- "Path to the 1st Entrance"
PathNum <- rbind(pathentr,numentr)
write.csv(PathNum, "../results/behaviordatasummary.csv", row.names = F)
```
