==> scripts/04_ephys.Rmd <==
---
title: "Electrophysiology Data Analysis"
author: "Rayna M Harris"
date: "November 23, 2017"
output: md_document
---

```{r setup, warning = F, message = F}
library(ggplot2)
library(reshape2)
library(tidyr) # for drop na function
library(cowplot)
library(dplyr)
library(magrittr)  # for function function "%<>%"

colorvalAPA2 <-  c( "#404040","#ca0020", "#bababa", "#f4a582")
```

Read and wrangle data

```{r data}
ephys2 <- read.csv("../data/Ephy_forDataAnalysisV2.csv", header=T, na.strings=c(""), stringsAsFactors = FALSE)

## rename columns 
names(ephys2)[names(ephys2)=="Peak..V."] <- "Peak V"
names(ephys2)[names(ephys2)=="training"] <- "group"

## new cols
ephys2$APA <- ifelse(grepl("TTT", ephys2$group), "consistent", 
                     ifelse(grepl("TCT", ephys2$group), "conflict",
                            ifelse(grepl("TTY", ephys2$group), "control",
                                   ifelse(grepl("TCY", ephys2$group), "control", NA))))

ephys2$APA2 <- ifelse(grepl("TTT", ephys2$group), "consistent", 
                      ifelse(grepl("TCT", ephys2$group), "conflict",
                             ifelse(grepl("TTY", ephys2$group), "yoked-consistent",
                                    ifelse(grepl("TCY", ephys2$group), "yoked-conflict", NA))))

ephys2$Year <- ifelse(grepl("151", ephys2$ID), "2015", 
                      ifelse(grepl("16-", ephys2$ID), "2016", NA))

names(ephys2)
## reorder, drop old group coloumn
ephys2 <- ephys2[c(2,15:18,3:14)]

### Drop the voltages with very little recording and the Vmax
ephys2 <- ephys2[c(-10,-12, -16, -17)]


## make characters either factors or numbers, as appropriate
colsfactor = c(1:5)
#colsnumeric = c(18)
ephys2[,colsfactor] %<>% lapply(function(x) as.factor(as.factor(x)))
#ephys2[,colsnumeric] %<>% lapply(function(x) as.numeric(as.integer(x)))

## make with max column with max i/o
ephys2$min <- apply(ephys2[c(7:13)], 1, min,na.rm=TRUE)
ephys2$max <- apply(ephys2[c(7:13)], 1, max,na.rm=TRUE)


## prep levels for visualization
ephys2$Genotype <- factor(ephys2$Genotype, 
                          levels = c("WT", "FMR1KO"))
ephys2$APA <- factor(ephys2$APA, 
                     levels = c("control", "consistent", "conflict"))

ephys2$APA2 <- factor(ephys2$APA2,
                      levels = c("yoked-consistent","consistent","yoked-conflict", "conflict"))


## make long and tidy
ephys2_long <- melt(ephys2, id = c(1:5))

## add numeric for stat smooth
ephys2_long$variablenumeric <- ifelse((ephys2_long$variable == "V0"), "1", 
                                      ifelse(grepl("V10", ephys2_long$variable ), "2",
                                             ifelse(grepl("V15", ephys2_long$variable ), "3",
                                                    ifelse(grepl("V20", ephys2_long$variable), "4", 
                                                           ifelse(grepl("V30", ephys2_long$variable), "5",
                                                                  ifelse(grepl("V40", ephys2_long$variable), "6",
                                                                         ifelse(grepl("V50", ephys2_long$variable), "7", NA)))))))
ephys2_long <- ephys2_long %>% drop_na()

ephys2_long$variablenumeric <- as.numeric(ephys2_long$variablenumeric)

ephys2_long$Genotype <- factor(ephys2_long$Genotype, 
                               levels = c("WT", "FMR1KO"))
ephys2_long$APA <- factor(ephys2_long$APA, 
                          levels = c("control", "consistent", "conflict"))




ephys2summaryNum <- dplyr::summarise(group_by(ephys2, Genotype, APA2), m = mean(min), se = sd(min)/sqrt(length(min)), len = length(min))
ephys2summaryNum <- as.data.frame(ephys2summaryNum)
levels(ephys2summaryNum$Genotype) <- c("WT","FMR1KO")

ephys2$genoAPA <- as.factor(paste(ephys2$Genotype,ephys2$APA2,sep="_"))
```

Plot of fEPSP slop by treatment group and genotype

```{r fEPSPslope}
# this was the third plot I made. Name left of from that
plot3 <- ephys2 %>% 
  ggplot(aes(x=Genotype, y=max, color=APA2)) +
  geom_violin(aes(fill = factor(Genotype))) +
  geom_point(size=2) +
  scale_y_continuous(trans = "reverse",
                     limits = c(0.001 , -0.002)) + 
  background_grid(major = "xy", minor = "none") + 
  theme_cowplot(font_size = 8, line_size = 0.25) + 
  scale_color_manual(values = colorvalAPA2) +
  scale_fill_manual(values=alpha(c( "#404040","white"), .1)) +
  labs(y = "Maximum fEPSP Slope", x = "Genotype") +
  facet_wrap(~APA2, nrow = 1) + 
  theme(legend.position="none")
plot3

pdf(file="~/Github/FMR1CA1rnaseq/figures/04_ephys/ephys3.pdf", width=6, height=2.5)
plot(plot3)
dev.off()
```

Statistics

```{r stats}
aov1 <- aov(min ~ Genotype, data=ephys2)
summary(aov1) 

aov1 <- aov(min ~ Genotype * APA2, data=ephys2)
summary(aov1) 
TukeyHSD(aov1)

aov1 <- aov(max ~ Genotype, data=ephys2)
summary(aov1) 
aov1 <- aov(max ~ Genotype * APA2, data=ephys2)
summary(aov1)

summary(ephys2$min)
```
